gh-27: Added time aware column definition
Some columns are not present for all the years. Examples:

 - Schutzstatus S: introduced in SSS 2024
 - Gemeinde BFS ID: introduced in SSS 2023

To ensure the correct parsing, added time aware column definitions
(class `ColumnConfig`). Made sure, that all columns over all years are
inserted, but with empty values for non-active years. This is important
for loading into the database.

This handles also the problem with the introduced gemeinde_bfs_id
columns, startin in 2023.