gh-37: Added groups to indicators
In multidimensional indicators, it does not make sense to insert every
dimensional value as a single indicator. This would lead to way too many
indicators. Instead, allow groups per indicator.

The plain total of an indicator is represented as `GROUP TOTAL` in the
value of group 1 to 4.

Additionally, implemented tests as jinja macro instead of making use of
YAML variables.