gh-37: Added indicators for Landwirtschaft
Added indicators for the Landwirtschaftliche Strukturerhebung.

Additionally, modified macro `intm_bfs_stat_tab`:
Before the change, if any `build_total_value` was set in the config of a
indicator in the mart, all indicators were treated as if they need a
calulation for the GROUP TOTAL. This is now determined for every
indicator individually.