gh-72: Added knowledge_date to response
Instead of returning valid_from/valid_to from the original snapshot
table. This should be enough and does persist the user's choice on the
knowledge date. This can be usefull, when the user is sharing the
retrieved data. With this approach, the user can get the exact same data
later again.

Additionally, fixed an error on period_ref, because the date was not
correctly parsed for SQL COPY. Now passing dates as ISO-8601 strings
which gets recognized by PSQL.