gh-77: Added Statatlas 2.0
The newly published Statatlas 2.0 has a complete overhaul of the data
presented to the user. For this reason (and because no API is available
yet), the import job has to be rewritten.

The modified jinja macro for `intm_bfs_statatlas` now merges the data
from Statatlas 1 with Statatlas 2.

From the viewpoint of a single indicator: Some indicator will be updated
with new data from Statatlas 2.0, others will remain untouched, until an
equivalent data source is found to replace the old data. Currently,
there are some indicators which will not update anymore until an
alternative is found.

Additionally, added a new column `_etl_version` to INTM models to keep
track, from which source version the data comes from.