gh-83: Fixed test for min count of municipalities
In INTM, the models get tested to make sure, that the amount of rows per
`period_ref` is at least 2000 (the expected amount of municipalities in
Switzerland).

Previously, this was done by grouping also by `knowledge_date_from`,
which is wrong. Sometimes, only the data of some municipalities
gets updated, which will fail the test `odapi_intm_unique_municipalities`.

Removed the gruping for `knowledge_date_from`, but now filter for only
the currently active rows to test against it.