Refactor INTM
Refactored INTM models with indicator data.

Previously, all INTM logic was done in one single model, with multiple
appended macros. If something went wrong, this structure was difficult
to debug.

The INTM models now have the following structure:

  1. intm_: First INTM model which brings the various data sources into
     a similar structure. All config to ODAPI are stored in the
     corresponding model config of this model. Downstream models read
     from this config.
  2. intm_*__measure: Calculates the measures as provided by the intm_
     model config.
  3. intm_*__typecast: Converts all columns into the desired data types,
     which makes it possible to then generate a partition table for
     mart_ogd_api
  4. mart_: Replaces text values for group, group_values and sources to
     IDs of the corrensponding DIM table.

A script has been written to automatically generate all intm_*__measure,
intm_*__typecast and mart_ models automatically, once the config and
modlel for intm_ has been written:

    create_intm_dunder_models.sh

Disk usage:
All INTM models are now materialized as views. Only the mart_ models
(which act as partition for mart_ogd_api) and the snapshot tables are
materialized as table. This helps to save storage in the future.

Additionally, data on mart_ gets normalized and replaced with IDs, so
that it takes less space while materializing as table.